app: add LibreVNA logo and use it as icon launcher
Now it's possible to identify LibreVNA with a logo. This logo would be
visible as an icon when app is launched.

A banner containing same logo has been added to README. Also,
a desktop launcher file is created to be used in GNU/Linux environments
enabling in this way the ability to add as favorite.

Closes #57